sdk-scripts/labs/_buildpacks/sample-py-app/requirements.txt
# https://pypi.org/project/flask/
Flask >= 2.2

# https://pypi.org/project/gunicorn/
gunicorn==20.1.0

google-cloud-translate

